chore: move legacy backend code to sub-dir
