+++ core/src/cloud/grow/api.ts
@@ -25,8 +25,8 @@ import { CloudApiError, GardenError } from "../../exceptions.js"
 import { gardenEnv } from "../../constants.js"
 import { LogLevel } from "../../logger/logger.js"
 import { getCloudDistributionName, getCloudLogSectionName } from "../util.js"
-import { getStoredAuthToken } from "../auth.js"
-import type { CloudApiFactoryParams, CloudApiParams } from "../api.js"
+import { getStoredAuthToken } from "../legacy/auth.js"
+import type { CloudApiFactoryParams, CloudApiParams } from "../legacy/api.js"
 import { deline } from "../../util/string.js"
 import { TRPCClientError } from "@trpc/client"
 import type { InferrableClientTypes } from "@trpc/server/unstable-core-do-not-import"
