+++ core/src/cloud/grow/auth.ts
@@ -12,9 +12,9 @@ import { getHTTPStatusCodeFromError } from "@trpc/server/http"
 import type { ClientAuthToken, GlobalConfigStore } from "../../config-store/global.js"
 import type { Log } from "../../logger/log-entry.js"
 import { describeTRPCClientError, getNonAuthenticatedApiClient } from "./trpc.js"
-import { CloudApiTokenRefreshError } from "../api.js"
+import { CloudApiTokenRefreshError } from "../legacy/api.js"
 import { CloudApiError } from "../../exceptions.js"
-import { clearAuthToken, saveAuthToken } from "../auth.js"
+import { clearAuthToken, saveAuthToken } from "../legacy/auth.js"
 import { getCloudDistributionName } from "../util.js"
 import dedent from "dedent"
 import { handleServerNotices } from "./notices.js"
