+++ core/src/cloud/legacy/auth.ts
@@ -14,16 +14,16 @@ import bodyParser from "koa-bodyparser"
 import Router from "koa-router"
 import getPort from "get-port"
 import cloneDeep from "fast-copy"
-import type { Log } from "../logger/log-entry.js"
-import { gardenEnv } from "../constants.js"
-import type { ClientAuthToken, GlobalConfigStore } from "../config-store/global.js"
-import { dedent } from "../util/string.js"
-import { GardenError, InternalError } from "../exceptions.js"
+import type { Log } from "../../logger/log-entry.js"
+import { gardenEnv } from "../../constants.js"
+import type { ClientAuthToken, GlobalConfigStore } from "../../config-store/global.js"
+import { dedent } from "../../util/string.js"
+import { GardenError, InternalError } from "../../exceptions.js"
 import { add } from "date-fns"
-import { getCloudDistributionName, getCloudLogSectionName } from "./util.js"
+import { getCloudDistributionName, getCloudLogSectionName } from "../util.js"
 import type { ParsedUrlQuery } from "node:querystring"
-import type { Garden } from "../index.js"
-import { styles } from "../logger/styles.js"
+import type { Garden } from "../../index.js"
+import { styles } from "../../logger/styles.js"
 
 class LoginRequiredWhenConnected extends GardenError {
   override type = "login-required"
