+++ core/src/cloud/legacy/workflow-lifecycle.ts
@@ -6,16 +6,16 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import type { WorkflowConfig } from "../config/workflow.js"
-import { makeRunConfig } from "../config/workflow.js"
-import type { Log } from "../logger/log-entry.js"
-import { CloudApiError } from "../exceptions.js"
-import { gardenEnv } from "../constants.js"
-import type { GardenWithOldBackend } from "../garden.js"
+import type { WorkflowConfig } from "../../config/workflow.js"
+import { makeRunConfig } from "../../config/workflow.js"
+import type { Log } from "../../logger/log-entry.js"
+import { CloudApiError } from "../../exceptions.js"
+import { gardenEnv } from "../../constants.js"
+import type { GardenWithOldBackend } from "../../garden.js"
 import type { ApiFetchResponse } from "./http-client.js"
 import type { CreateWorkflowRunResponse } from "@garden-io/platform-api-types"
-import { dedent } from "../util/string.js"
-import { GotHttpError } from "../util/http.js"
+import { dedent } from "../../util/string.js"
+import { GotHttpError } from "../../util/http.js"
 
 export interface RegisterWorkflowRunParams {
   workflowConfig: WorkflowConfig
