+++ core/src/commands/cloud/secrets/secrets-create.ts
@@ -16,7 +16,7 @@ import { PathParameter, StringParameter, StringsParameter } from "../../../cli/p
 import type { SecretResult } from "./secret-helpers.js"
 import { makeSecretFromResponse } from "./secret-helpers.js"
 import { getEnvironmentByNameOrThrow } from "./secret-helpers.js"
-import type { Secret } from "../../../cloud/api.js"
+import type { Secret } from "../../../cloud/legacy/api.js"
 import { handleSecretsUnavailableInNewBackend } from "../../../cloud/secrets.js"
 
 export const secretsCreateArgs = {
