+++ core/src/commands/login.ts
@@ -12,8 +12,8 @@ import { printHeader } from "../logger/util.js"
 import dedent from "dedent"
 import type { Log } from "../logger/log-entry.js"
 import { CloudApiError, ConfigurationError, InternalError, TimeoutError } from "../exceptions.js"
-import type { AuthToken } from "../cloud/auth.js"
-import { AuthRedirectServer, saveAuthToken } from "../cloud/auth.js"
+import type { AuthToken } from "../cloud/legacy/auth.js"
+import { AuthRedirectServer, saveAuthToken } from "../cloud/legacy/auth.js"
 import type { EventBus } from "../events/events.js"
 import { getCloudDomain } from "../cloud/util.js"
 import type { GardenBackend } from "../cloud/backend.js"
