+++ core/src/commands/workflow.ts
@@ -25,7 +25,7 @@ import fsExtra from "fs-extra"
 import { getDurationMsec, toEnvVars } from "../util/util.js"
 import { runScript } from "../util/util.js"
 import { LogLevel } from "../logger/logger.js"
-import { registerWorkflowRun } from "../cloud/workflow-lifecycle.js"
+import { registerWorkflowRun } from "../cloud/legacy/workflow-lifecycle.js"
 import { parseCliArgs, pickCommand, processCliArgs } from "../cli/helpers.js"
 import type { GlobalOptions, ParameterValues } from "../cli/params.js"
 import { StringParameter } from "../cli/params.js"
