+++ core/src/events/events.ts
@@ -8,7 +8,7 @@
 
 import { omit } from "lodash-es"
 import EventEmitter2 from "eventemitter2"
-import type { LogEntryEventPayload } from "../cloud/restful-event-stream.js"
+import type { LogEntryEventPayload } from "../cloud/legacy/restful-event-stream.js"
 import type { DeployStatusForEventPayload } from "../types/service.js"
 import type { RunStatusForEventPayload } from "../plugin/base.js"
 import type { Omit, PickFromUnion } from "../util/util.js"
@@ -17,7 +17,7 @@ import type { CommandInfo, EventNamespaceStatus } from "../plugin-context.js"
 import type { GraphResult } from "../graph/results.js"
 import type { BuildStatusForEventPayload } from "../plugin/handlers/Build/get-status.js"
 import type { ActionStatusPayload } from "./action-status-events.js"
-import type { AuthToken } from "../cloud/auth.js"
+import type { AuthToken } from "../cloud/legacy/auth.js"
 
 interface EventContext {
   gardenKey?: string
