+++ core/src/logger/logger.ts
@@ -13,7 +13,7 @@ import { CommandError, ParameterError, InternalError } from "../exceptions.js"
 import { TerminalWriter } from "./writers/terminal-writer.js"
 import { JsonTerminalWriter } from "./writers/json-terminal-writer.js"
 import { EventBus } from "../events/events.js"
-import { formatLogEntryForEventStream } from "../cloud/restful-event-stream.js"
+import { formatLogEntryForEventStream } from "../cloud/legacy/restful-event-stream.js"
 import { gardenEnv } from "../constants.js"
 import { getEnumKeys } from "../util/util.js"
 import { range } from "lodash-es"
