+++ core/src/plugins/container/build.ts
@@ -27,7 +27,7 @@ import type { ActionLog } from "../../logger/log-entry.js"
 import type { PluginContext } from "../../plugin-context.js"
 import { cloudBuilder } from "./cloudbuilder.js"
 import { styles } from "../../logger/styles.js"
-import type { CloudBuilderAvailableV2 } from "../../cloud/api.js"
+import type { CloudBuilderAvailableV2 } from "../../cloud/legacy/api.js"
 import { renderTimeDurationMs, spawn, type SpawnOutput } from "../../util/util.js"
 import { isSecret, type Secret } from "../../util/secrets.js"
 import { tmpdir } from "os"
