+++ core/src/server/instance-manager.ts
@@ -11,7 +11,7 @@ import type { AutocompleteSuggestion } from "../cli/autocomplete.js"
 import { Autocompleter } from "../cli/autocomplete.js"
 import { parseCliVarFlags } from "../cli/helpers.js"
 import type { ParameterObject, ParameterValues } from "../cli/params.js"
-import type { GardenCloudApi } from "../cloud/api.js"
+import type { GardenCloudApi } from "../cloud/legacy/api.js"
 import type { BuiltinArgs, Command } from "../commands/base.js"
 import { getBuiltinCommands, flattenCommands } from "../commands/commands.js"
 import { getCustomCommands } from "../commands/custom.js"
