+++ core/src/server/server.ts
@@ -28,8 +28,8 @@ import { EventBus, shouldStreamWsEvent } from "../events/events.js"
 import type { ValueOf } from "../util/util.js"
 import { joi } from "../config/common.js"
 import { dedent, randomString } from "../util/string.js"
-import { authTokenHeader } from "../cloud/auth.js"
-import type { ApiEventBatch, LogEntryEventPayload } from "../cloud/restful-event-stream.js"
+import { authTokenHeader } from "../cloud/legacy/auth.js"
+import type { ApiEventBatch, LogEntryEventPayload } from "../cloud/legacy/restful-event-stream.js"
 import { LogLevel } from "../logger/logger.js"
 import { EventEmitter } from "eventemitter3"
 import { sanitizeValue } from "../util/logging.js"
