+++ core/src/util/testing.ts
@@ -53,7 +53,7 @@ import { createReadStream, createWriteStream } from "fs"
 import got from "got"
 import { createHash } from "node:crypto"
 import { pipeline } from "node:stream/promises"
-import type { GardenCloudApiFactory } from "../cloud/api.js"
+import type { GardenCloudApiFactory } from "../cloud/legacy/api.js"
 import { parseTemplateCollection } from "../template/templated-collections.js"
 import type { VariablesContext } from "../config/template-contexts/variables.js"
 import { uuidv4 } from "./random.js"
