+++ core/test/unit/src/commands/cloud/secrets/secrets-update.ts
@@ -15,7 +15,7 @@ import {
 } from "../../../../../../src/commands/cloud/secrets/secrets-update.js"
 import { deline } from "../../../../../../src/util/string.js"
 import { createProjectConfig, expectError, getDataDir, makeTestGarden } from "../../../../../helpers.js"
-import type { Secret, SingleUpdateSecretRequest } from "../../../../../../src/cloud/api.js"
+import type { Secret, SingleUpdateSecretRequest } from "../../../../../../src/cloud/legacy/api.js"
 import { FakeGardenCloudApi } from "../../../../../helpers/api.js"
 
 describe("SecretsUpdateCommand", () => {
