+++ core/test/unit/src/commands/login.ts
@@ -10,11 +10,11 @@ import { expect } from "chai"
 import * as td from "testdouble"
 import type { TempDirectory } from "../../../helpers.js"
 import { expectError, getDataDir, makeTempDir, makeTestGarden, withDefaultGlobalOpts } from "../../../helpers.js"
-import { AuthRedirectServer, getStoredAuthToken, saveAuthToken } from "../../../../src/cloud/auth.js"
+import { AuthRedirectServer, getStoredAuthToken, saveAuthToken } from "../../../../src/cloud/legacy/auth.js"
 
 import { LoginCommand, rewriteProjectConfigYaml } from "../../../../src/commands/login.js"
 import { randomString } from "../../../../src/util/string.js"
-import { GardenCloudApi } from "../../../../src/cloud/api.js"
+import { GardenCloudApi } from "../../../../src/cloud/legacy/api.js"
 import { LogLevel } from "../../../../src/logger/logger.js"
 import { gardenEnv } from "../../../../src/constants.js"
 import { getLogMessages } from "../../../../src/util/testing.js"
