+++ core/test/unit/src/commands/logout.ts
@@ -11,7 +11,7 @@ import * as td from "testdouble"
 import type { TempDirectory } from "../../../helpers.js"
 import { getDataDir, makeTempDir, makeTestGarden, withDefaultGlobalOpts } from "../../../helpers.js"
 import { randomString } from "../../../../src/util/string.js"
-import { GardenCloudApi } from "../../../../src/cloud/api.js"
+import { GardenCloudApi } from "../../../../src/cloud/legacy/api.js"
 import { LogLevel } from "../../../../src/logger/logger.js"
 import { LogOutCommand } from "../../../../src/commands/logout.js"
 import { expectError, getLogMessages } from "../../../../src/util/testing.js"
@@ -20,7 +20,7 @@ import { DEFAULT_GARDEN_CLOUD_DOMAIN } from "../../../../src/constants.js"
 import { GlobalConfigStore } from "../../../../src/config-store/global.js"
 import type { Garden } from "../../../../src/index.js"
 import { makeDummyGarden } from "../../../../src/garden.js"
-import { getStoredAuthToken, saveAuthToken } from "../../../../src/cloud/auth.js"
+import { getStoredAuthToken, saveAuthToken } from "../../../../src/cloud/legacy/auth.js"
 import { uuidv4 } from "../../../../src/util/random.js"
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
