+++ core/test/unit/src/garden.ts
@@ -67,7 +67,7 @@ import { TreeCache } from "../../../src/cache.js"
 import { omitUndefined } from "../../../src/util/objects.js"
 import { add } from "date-fns"
 import stripAnsi from "strip-ansi"
-import { GardenCloudApi } from "../../../src/cloud/api.js"
+import { GardenCloudApi } from "../../../src/cloud/legacy/api.js"
 import { GlobalConfigStore } from "../../../src/config-store/global.js"
 import { getRootLogger } from "../../../src/logger/logger.js"
 import { uuidv4 } from "../../../src/util/random.js"
