+++ core/test/unit/src/server/server.ts
@@ -11,7 +11,7 @@ import type { Server } from "http"
 import { GardenServer, startServer } from "../../../../src/server/server.js"
 import type { Garden } from "../../../../src/garden.js"
 import { expect } from "chai"
-import { authTokenHeader } from "../../../../src/cloud/auth.js"
+import { authTokenHeader } from "../../../../src/cloud/legacy/auth.js"
 import { ServeCommand } from "../../../../src/commands/serve.js"
 import { gardenEnv } from "../../../../src/constants.js"
 import { deepOmitUndefined } from "../../../../src/util/objects.js"
